Add and remove items from cart.

+++ HoneyQBE/src/main/java/be/honeyq/HoneyQBE/repository/ArticleRepository.java
@@ -3,4 +3,4 @@ package be.honeyq.HoneyQBE.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import be.honeyq.HoneyQBE.model.Article;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {}
+public interface ArticleRepository extends JpaRepository<Article, Long> {}
